Add util func for processing zvm-image args
The user can specify multiple --zvm-image arguments on the command line;
each image can include optional mount point and access (ro or rw)
parameters, which needs to be written to the fstab in the nvram file.

This function processes the command line arguments into something which
can easily be written into the nvram file.