Add some helper funcs for handling runtime files
`_get_runtime_file_paths` is used to generate the paths to the nvram,
manifest, boot, stdout, and stderr files (based on the working
directory).

`_check_runtime_files` checks if any of these files already exists and
raises a RuntimeError if any do exist. This is necessary because zvsh
needs to write these files to the specified working dir (which can be a
temp dir) and should not overwrite files without warning.