zvsh: Add `run_zerovm` function
This replaces the old implementation of ZvShell. This is still a work in
progress and there are still tests needed. As of now, most of the file
generation is well tested, but the entire workflow is not. We need some
good integration tests for this.

The gdb path through zvsh is unaffected by this change.