Remove inappropriate os.path.expandpath calls
This was prompted by 5c63955 which added an os.path.expandpath call to
the argument to --zvm-image.

I don't believe that we need to expand '~' and '~foo' in the command
line arguments to our scripts. That is the job of the shell and
normally works just fine with no effort on our part:

  % python -c 'import sys; print sys.argv' --some-arg ~/foo.txt
  ['-c', '--some-arg', '/home/mg/foo.txt']

The only problem is if users specify paths without a space:

  % python -c 'import sys; print sys.argv' --some-arg ~/foo.txt
  ['-c', '--some-arg=~/foo.txt']

If one insists on using no space, the correct way is to use an
environment variable:

  % python -c 'import sys; print sys.argv' --some-arg=$HOME/foo.txt
  ['-c', '--some-arg', '/home/mg/foo.txt']

This is consistent with how other tools such as sed and git works:

  % echo 's/foo/bar/' > ~/test.sed
  % echo hello foo | sed --file=~/test.sed
  sed: couldn't open file ~/test.sed: No such file or directory

  % echo 'commit msg' > ~/msg.txt
  % git commit --file=~/msg.txt
  fatal: could not read log file '~/msg.txt': No such file or directory

Paths read from configuration files (where there is no shell involved)
can be run through os.path.expanduser.