zvsh: escape nvram values more consistently
I dug into the ZRT config parser code and believe this should be a
more correct and complete way to escape the values we put into the
nvram files.

Before, a command line argument like '"quoted"' would have been put
into the nvram file as

  args = "quoted"

and this is wrong: the outer quotes are removed by ZRT when is parses
the nvram file. We now correctly escape double-quotes as well as
space, comma, backslash, and newline.