bundle: generate job description template (fixes #29)
The template is generated by 'zpm bundle'. For a zar called foo.zar,
the job description is included as foo.json in the root of the
tarball. Later, we might want to create a reserved directory for such
meta data.

To keep 'zpm deploy --execute' working, it will now load the job
description from the zar instead of generating it. The job description
template is then filled in with the Swift URL of the zar before the
job is executed.