zpm: update and correct `create_project`
The `create_project` was old and wrong. Previously, it assumed the
creation of a handful of special directories. Now, all we need to do is
create a default `zar.json` file in the target directory.

TODO: I took my best guess at what the default zar.json contents should
be. This needs revision and more documentation.