commands: `new` command -- catch and print errors
If, for example, the target `zar.json` file already exists, catch the
exception and print a nice error message for the user (instead of the
entire stack trace).

TODO: Instead of using RuntimeError, it might be better to use a custom
exception class, so that we don't accidentally mask other unexpected
RuntimeErrors.