zpm,commands: create_project returns zar.json path
The `zpm.create_project` function now return the full zar.json path,
which allows the command line interface to notify the user that the file
was created. We now print the full path, instead of just telling the
user into which directory the file was created. I think this is more
useful and explicit.