tox: run cov report with py27 and py3 configs
Instead of running a separate coverage report (using only python2.7,
consequently), this change forces command-line coverage reports to be
run with py27 and py3. This is useful because the coverage numbers can
vary from version to version, given the need for try/except for
importing stdlib stuff. This is also helpful in case we need to use
something like `six`, which could introduce python version specific
branches in logic.