new: keep ordering in generated zar.json
Before, we just kept the zar.json template in a normal dict. Since a
dict has no ordering, the generated zar.json did not look like the
input dict.

We now keep the template as an OrderedDict. That way we can present
the keys in a "nicer" order and ensure that the generated zar.json
looks more like our documentation.