ui: add support for input/output devices
When the zar.json file has an input or output device for a group,
there will now be a corresponding input field. The path entered there
is a relative Swift path, meaning that it will be used to construct a
Swift URL of the form

  swift://<user>/<relative path>

That way, the user doesn't have to care about the 'AUTH_f7e9...' part
of the URL. Later, we can replace this with a JavaScript file picker
that lets the user browse the Swift objects.