setup: include templates as package_data
The templates were not packaged and thus also not installed. This
meant that running 'zpm new' failed with:

  jinja2.exceptions.TemplateNotFound: zar.yaml

The error only occurs if the zpm package is installed normally,
creating a link with 'pip install -e .' hides the error.

Versions of Python before 2.7 did not automatically included the files
listed in package_data in the MANIFEST. So we now list them there too.