zpm: add auth version to deploy options
This allows you to deploy using both v1 and v2 auth.

We pass the data into the template as an escaped JSON string. This is
important to ensure proper escaping. Strings (password in particular)
can contain tricky characters such as quotes or backslashes and they
need to be escaped correctly.