deploy: standardize default in cmdline help texts
We now use %(default)s where possible to let argparse insert the
correct default in the help text.

The flags that default to an environment variable list the environment
variable instead of the actual value -- listing the default would not
be helpful when the environment variable isn't set (it would be the
empty string).