tox: avoid environment values bleeding into docs (fixes #101)
The defaults for the command line flags are shown in the generated
documentation. When the default comes from an environment variable,
this can show unwanted information.

We now set _TOX_SPHINX when running sphinx from tox and the arg
decorator will use this as a signal to avoid setting the default
value.