zpm: warn if a bundling/ui glob matches nothing
In the zapp.yaml, files can be selected for bundling by specifying lists
of files or glob patterns in the "bundling" or "ui" section.

If a file doesn't exist or if the pattern matches nothing, we now log a
warning. This probably means the user has done something wrong.

Cram tests are also updated, with warnings printed to the console for
missing files/bad globs.