zpm: raise an error if no files were added to tar
If no files were added to the zapp/tar (aside from the zapp.yaml),
probably it means that the "bundling" or "ui" patterns are wrong in some
way. This results in basically an empty zapp, which is not very useful.

This involved refactoring the bundling code a bit, and the ordering of
file adds has changed. This warranted updating cram tests as well.