tox: drop support for Python 3.2
Python version 3.0 to 3.2 had no support for u'...' Unicode literals.
The support was added again in Python 3.3 to make it easier to port
code from Python 2.x.

The lack of u'...' support has not been a direct problem for us, we
would like to use python-swiftclient instead of our own miniswift
module (#31). However, python-swiftclient only supports Python 3.3 and
later. More precisely, python-swiftclient tries to import simplejson
and this fails with a SyntaxError on Python 3.2.

The simplejson project will probably not support Python 3.2 anytime
soon:

  https://github.com/simplejson/simplejson/issues/66

We will therefore also drop support for Python 3.2 for now. We will
work with the upstream projects to see if it is easy to restore Python
3.2 support. The first pull request for one of our dependencies has
been positively received:

  https://github.com/kennethreitz/requests/pull/2064

Pip is also affected by this problem and an issue has been created:

  https://github.com/pypa/pip/issues/1839