zpm: (deploy) prepare authentication args first
This just cleans up the order in which things happen during deployment.
Previously, it was:

1. Get connection
2. Upload .zapp file
3. Upload job .json file
4. Prepare auth args
5. Render UI (with auth args), upload UI

The goal is to group the uploads together better, so that uploads all
happen together (and could be abstracted into a separate function). The
new order is as follows:

1. Get connection
2. Prepare auth args
3. Upload .zapp file
4. Upload job .json file
5. Render UI (with auth args), upload UI

No functionality changes here; this is pure refactoring.