zpm: do preauth in `deploy_project`
Before we use the ZeroCloudConnection to deploy files, a preauth is
called. This is meant to ensure that the connection is always in a
consistent state before the actual deployment begins.

This helps with refactoring, too. For example, while trying to
refactor/reorder a bit of code that used this connection, I found that
the preparation of one upload relied on the fact that the previous
upload succeeded and set the `url` attributes of the connection object.

Doing this explicitly up-front prevents this sort issue from happening.