deploy: show full URL to the deployed index.html (fixes #134)
Before, we would show the user that we had deployed into

  server/v1/account/container/path/

This relied on the staticweb middleware being installed and a suitable
web index being configured. We now show

  server/v1/account/container/path/index.html

instead when an index.html file is being deployed. If no index.html
file is uploaded (zapps with a non-standard UI), we show the same URL
as before and the users will have to sort things out by themselves.