zpm: remove Deployer class
Classes typically represent things -- they are normally named with
nouns to reflect this. The Deployer class was encapsulating a
behavior, the deployment of a zapp. This was reflected in the fact
that the class had just one public method, this was the class' sole
behavior. The state of the class was not very interesting.

The class thus behaved very much like a function. This is a very
direct translation of the class and its methods into a set of
functions.