zpm: inline single use _upload function
When adapting the tests, it was discovered that the previous mocking
was wrong. The zpm._prepare_upload function was mocked to return

  [('a', 'b'), ('c', 'd')]

even though this is an impossible return value for it -- the function
will always return tuples where the first element contains a '/'. This
is the slash that separates the container name from the object name.

This was not discovered earlier because zpm._uploads was mocked too in
the tests and so it ignored its arguments.