zpm: rename a variable in `deploy_project`
Instead of `version`, the variable is renamed to `ui_auth_version`. The
difference here is subtle, because the auth version is used in two ways
in zpm:

1) To authenticate to Swift/ZeroCloud services
2) To populate auth information in UI templates of zapps (at deploy
   time)

The first is determined within the `_get_zerocloud_conn`. The second can
also be determined within `_get_zerocloud_conn`, but can be overridden
by the `--no-ui-auth` command line arg.