Add project type-specific bundling (python)
This change adds an extensible but simple framework for bundling
specific types of projects. The default and only bundler implemented
currently is `python`. (This was implicitly true anyway before, so now
it's explicit.)

Includes adding a `--refresh-deps` option to `zpm bundle`.
For the python project bundling template, if --refresh-deps/-r flag is
`True` we recreate the tox-managed venv from which we bundle 3rd party
python dependencies.