zpm: use zapptemplate to create proj template
This removes a lot of the complexity from the `zpm` module and relies
instead on the `zapptemplate` module for doing most of the work in
creating zapp.yaml and other project files.

This also integrates the `zpm new --template <template>` switch in the
project creation code.

Tests are modified to reflect this change. Some obsolete tests are
deleted.