Move scripts into scripts/ directory
This is a bit cleaner than putting scripts like zvm, zvsh, etc. in the
root directory of the repo. For one, that means less clutter in the
root. Two, it also means that we don't have to name packages/folders
weird things like zvshlib, zvmlib, and zpmlib; we can just call them
zvsh, zvm, or zpm. That's much nicer.