debian: remove explicit `zvsh` script install
In commit 2789559af309e63736cf35e74c72c22bf6618dd4, all scripts (`zvsh`,
`zpm`, `zvsh`) were moved into a `scripts/` directory in the root of the
repo. This broke packaging, because the `debian/zerovm-cli.install` file
was expecting the `zvsh` script to still be in the root of the repo,
instead of in the `scripts/` directory.

In fact, this directive wasn't needed; the `setup.py` and `debian/rules`
take care of script placement anyway.