zpm: properly set the exec group name on bundle
There are some subtle nuances with how the nexe name and the args are
interpreted. Settings the ['exec']['name'] property explcitly in the
system.map helps to sidestep these issues and prevent confusion.

In the case of a exec group path of `file://python2.7:python`, the
['exec']['name'] is set to `python`. Similarly (and more importantly),
for a case like `swift://~/some_container/busybox`, the ['exec']['name']
will be set to `busybox`. In the second especially, we need to
explicitly set the exec name because otherwise, zerocloud will interpret
the execution group name (from the zapp.yaml) as argv[0], which causes
weird bugs.

This patch helps make that whole situtation less volatile by making the
system.map more explicitly defined.